requirements.txt
Django>=5.0,<6.1
psycopg2-binary>=2.9
whitenoise>=6.7
django-versatileimagefield>=2.0
gunicorn>=22.0
python-dotenv
dj-database-url